simplified C++ sample added, Lua, AS samples updated
